Serial - CDCACM is starting to work!
I have tested it talking to a Teensy 3.2, both simple program and emulating an AX Servo controller.

I ran the T3.2 with my servo controller code and it properly received query for local registers and returned a value.

I tested it for whole device (Serial only) and then with Serial+Joystick+... and it appeared to also work with the Interface only mode.